Adds Backbones to FRCNN Take 2 (#475)
* Adds frcnn

* update tests

* aaply suggestions from review

* refactor later

* leave any

* Apply suggestions from code review


* remove unused imports

* black, isort, cleanup

* fix init issue

* noqa

* Fix isort

* black format and refacotor

* Fix isort again

* Fix isort

* decouple sutff

* yapf formatting

* Yapf reformat

* reformat

* converts warnings

* Formatted

* Fix bug

* Apply suggestions from code review


* format yapf

* refactor

* chlog

+++ pl_bolts/datasets/dummy_dataset.py
@@ -78,8 +78,8 @@ class DummyDetectionDataset(Dataset):
 
     def __getitem__(self, idx: int):
         img = torch.rand(self.img_shape)
-        boxes = torch.tensor([self._random_bbox() for _ in range(self.num_boxes)])
-        labels = torch.randint(self.num_classes, (self.num_boxes, ))
+        boxes = torch.tensor([self._random_bbox() for _ in range(self.num_boxes)], dtype=torch.float32)
+        labels = torch.randint(self.num_classes, (self.num_boxes, ), dtype=torch.long)
         return img, {"boxes": boxes, "labels": labels}
